Some major updates, converted to use our latest methods!
Clean and lightweight starter theme based on Bootstrap 5, jQuery 3, SCSS and incorporates loads of functions ready to be
turned on and off in the functions.php file.

This theme strips WordPress back, removes all the bloat and leaves you with a clean html response, a favicon, a single
css resource in the head and a single js resource in the footer. Four requests in total.

Some features within this theme may not be laid out the way you like to work, but feel free to customise it. I developed
this theme for ourselves and to streamline getting started on a new bespoke theme for a project.

The JS and SCSS within this theme are compiled using Vite from the /src folder to the /build folder. A good
understanding of npm and composer is required to use this theme.

Feel free to submit pull requests of any improvements you can make; and if you find this theme useful, please
[buy me a coffee](https://www.buymeacoffee.com/mrl22).